Refactored property helpers into multiple files
